Refactor: add DTO (data transfer object) class. This DTO model is used to transfer data from client to server or from server to client.
The original APIs receive or return a Customer object , which is part of the domain model (considered as implementation details, which can change frequently, and these changes can break existing clients that depend on the Customer object). Therefore we need to make the contract of these apis as stable as possible.

Secondly, by receiving or returning domain objects, we are opening a security hole. Hacker can easily pass additional data into json and it will be mapped to domain object, even for properties that should not be updated. By using a DTO, you can simply exclude the properties that willnot be updated.